Retype OpType from c_uint to c_int and drop its 150 casts
The `OP_*` ids have exactly one consumer -- `oparg_T::op_type`, a
`c_int` -- but c2rust typed the anonymous enum they came from as the
`c_uint` the C compiler happened to pick for it. Every one of the 150
use sites in the tree therefore spelled itself `OP_X as c_int`, and
`get_op_type`, `get_op_char` and friends all take and answer `c_int`.

Retyping the alias and deleting the casts is mechanical (a regex over
`\b(OP_\w+)\s+as\s+(::core::ffi::)?c_int\b`, 25 files) and has to happen
in one commit: with `OpType = c_int` a surviving cast is
`clippy::unnecessary_cast`, which would be growth in the twelve files
outside this slice that hold one. This is the `CMOD_*` situation the
ratchet traps describe, resolved rather than deferred.

Behaviour-neutral: the values are 0..=29 and the two spellings agree on
every one of them. opsweep byte-identical.

+++ crates/nvim/src/nvim/eval/funcs/region.rs
@@ -340,7 +340,7 @@ unsafe fn block_oparg(
         oparg_T {
             motion_type: kMTBlockWise,
             inclusive: true,
-            op_type: OP_NOP as c_int,
+            op_type: OP_NOP,
             start: p1,
             end: p2,
             start_vcol,
